Add some common allocation strategy tests
